requirements.txt
absl-py==2.1.0
albumentations==1.4.6
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
clang==5.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
easydict==1.13
filelock==3.14.0
fire==0.6.0
Flask==3.0.3
flatbuffers==1.12
fonttools==4.51.0
fsspec==2024.3.1
gast==0.4.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.63.0
gunicorn==22.0.0
h5py==3.1.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.42.0
logger==1.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
mtcnn==0.1.1
namex==0.0.8
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.17.3
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.1
pooch==1.8.1
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rembg==2.0.56
requests==2.31.0
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.15.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.34.0
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.5.10
timm==0.9.16
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchaudio==0.9.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.13
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.3
wget==3.2
wrapt==1.12.1
zipp==3.18.1
